+++ src/recordSwap/recordSwap.cpp
@@ -1,6 +1,6 @@
 /*
  * Algorithm for targeted record swapping
- * Version: 0.3.1
+ * Version: 0.3.2
  */
 
 #include <iostream>